requirements.txt
async-timeout==5.0.0
beelib @ git+https://github.com/BeeGroup-cimne/beelib@5623052fcfee079dd73802fe2d021a6df6fce175
beemeteo @ git+https://github.com/BeeGroup-cimne/beemeteo@18dc622d58ad83145cabff200726f70b16861abc
cachetools==5.5.0
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
duckdb==0.10.2
durationpy==0.9
elementpath==4.4.0
falcon==3.1.3
fonttools==4.53.1
google-auth==2.37.0
greenlet==3.0.3
h3==4.1.2
happybase==1.2.0
idna==3.7
influxdb-client==1.47.0
isodate==0.6.1
Jinja2==3.1.4
jsonpath-python==1.0.6
kafka-python==2.0.2
kaleido==0.2.1
kiwisolver==1.4.5
kubernetes==31.0.0
load-dotenv==0.1.0
MarkupSafe==3.0.2
matplotlib==3.9.1
morph_kgc==2.7.0
neo4j==5.19.0
numpy==1.26.4
oauthlib==3.2.2
packaging==24.1
pandas==2.2.2
pillow==10.4.0
plotly==5.24.1
ply==3.11
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.20.0
pydruid==0.6.6
pymodbus==3.6.8
pyoxigraph==0.3.22
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-forecastio==1.4.0
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
rdflib==7.0.0
reactivex==4.0.4
redis==5.2.0
requests==2.31.0
requests-oauthlib==2.0.0
responses==0.25.3
rsa==4.9
ruamel.yaml==0.18.0
ruamel.yaml.clib==0.2.8
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.29
tenacity==9.0.0
thriftpy2==0.4.20
timezonefinder==6.5.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
websocket-client==1.8.0
xarray==2024.10.0
